requirements.txt
streamlit
google-generativeai
googlemaps
python-dotenv
pillow
requests
pydeck
pandas
